fix(eval): eval-harness stops hand-rolling a narrow event shape
src/eval/eval-harness.ts built its own event per sample:

    { type: sample.eventType, content: sample.text, fields: sample.fields }

It is an FP/TP harness, so it was supposed to go through the canonical
shape set — but tests/corpus-event-shapes.test.ts only ever ratcheted
files under scripts/, and this one lives under src/. It is the harness
behind `npm run eval` and `npm run eval:pint`, i.e. the numbers that
reach README badges, data/measurements/ and the website.

The cost of the private shape: every PINT sample is typed llm_input,
src/engine.ts maps that to source llm_io, and the source-type filter
then skips every rule whose agent_source.type differs. On this base that
is 383 of the 773 non-draft rules. They could not fire, so they could
not be counted as false positives either — the published PINT
"precision 100.0%" was measured over roughly half the rulebase.

Changes:
  - move the canonical shape set to src/corpus-event.ts so src/ can
    import it (tsconfig rootDir is ./src). scripts/lib/corpus-event.ts
    stays as a re-export, so every script import and the ratchet's grep
    are untouched.
  - add sourceAgnosticEvent(): the same event re-typed to mcp_exchange,
    which removes ONLY the source-type filter. Lets a caller separate
    rule admission from field pouring.
  - eval-harness takes an explicit EvalEventShapeMode
    (legacy | wide | canonical), default canonical, override with
    ATR_EVAL_EVENT_SHAPE. legacy reproduces every previously published
    number byte for byte. The mode is printed by the runners and written
    into the measurement notes, so no measurement file can be ambiguous
    about which presentation produced it.
  - the Tier 2.5 embedding pass still runs exactly once per sample, on
    the declared event: it is not source-filtered and reads only
    `content`, so re-encoding the same text four more times would cost
    5x runtime for zero verdict change.
  - add src/eval/eval-harness.ts to the ratchet so this cannot recur.

No published number file is touched by this commit. The before/after
measurements are reported to the human for a decision.
