+++ src/eval/eval-harness.ts
@@ -16,6 +16,7 @@ import { resolve, join } from 'node:path';
 import { existsSync, readFileSync, writeFileSync } from 'node:fs';
 import { writeMeasurementFromEvalReport } from '../measurement/from-eval-harness.js';
 import { ATREngine } from '../engine.js';
+import { corpusShapes, sourceAgnosticEvent } from '../corpus-event.js';
 import { EmbeddingModule } from '../modules/embedding.js';
 import type { AgentEvent } from '../types.js';
 import type { CorpusSample } from './corpus.js';
@@ -24,6 +25,46 @@ import type { SampleResult, EvalReport, BaselineThresholds, RegressionCheck } fr
 import { computeEvalReport, checkRegression } from './metrics.js';
 import type { RuleQualityReport } from './rule-metrics.js';
 import { computeRuleQuality } from './rule-metrics.js';
+
+/**
+ * How a corpus sample is presented to the engine.
+ *
+ *   'legacy'    — one event: `{ type: sample.eventType, content: sample.text }`.
+ *                 What this harness did until 2026-08-05. Reproduces every
+ *                 previously published eval / PINT number.
+ *   'wide'      — legacy PLUS the same event re-typed to mcp_exchange, which
+ *                 removes ONLY the engine's source-type filter. Isolates how
+ *                 much of the gap is rule ADMISSION rather than field pouring.
+ *   'canonical' — legacy PLUS scripts' canonical shape set (src/corpus-event.ts)
+ *                 PLUS scanSkill(). The presentation every FP gate in this repo
+ *                 already uses.
+ *
+ * WHY THIS EXISTS AT ALL
+ *
+ * src/engine.ts derives a source type from event.type (EVENT_TYPE_TO_SOURCE)
+ * and then skips every rule whose `agent_source.type` differs. The PINT corpus
+ * labels all 850 samples `llm_input` -> source `llm_io`, so on main 383 of the
+ * 773 non-draft rules — every mcp_exchange, tool_call, multi_agent_comm,
+ * memory_access, context_window and skill_lifecycle rule — were dropped before
+ * a single pattern was compared. They could not produce a true positive and,
+ * more importantly, could not produce a false positive either: the published
+ * "precision 100%" was measured over half a rulebase.
+ *
+ * That is the same zero-measurement defect src/corpus-event.ts was written to
+ * abolish; this harness simply escaped the ratchet because it lives under src/
+ * and the ratchet only greps scripts/.
+ *
+ * The mode is explicit, and recorded in the measurement notes, because flipping
+ * it changes a PUBLISHED number. Override with ATR_EVAL_EVENT_SHAPE=legacy|wide.
+ */
+export type EvalEventShapeMode = 'legacy' | 'wide' | 'canonical';
+
+/** Default presentation. Env override exists so an old number can be reproduced. */
+export function defaultEventShapeMode(): EvalEventShapeMode {
+  const raw = process.env['ATR_EVAL_EVENT_SHAPE']?.trim().toLowerCase();
+  if (raw === 'legacy' || raw === 'wide' || raw === 'canonical') return raw;
+  return 'canonical';
+}
 
 export interface EvalConfig {
   /** Directory containing ATR YAML rules */
@@ -38,12 +79,19 @@ export interface EvalConfig {
   readonly outputPath?: string;
   /** Enable Tier 2.5 embedding evaluation (default: auto-detect) */
   readonly enableEmbedding?: boolean;
+  /** Event presentation (default: defaultEventShapeMode()) */
+  readonly eventShape?: EvalEventShapeMode;
 }
 
 /**
- * Convert a corpus sample to an AgentEvent.
+ * The sample exactly as its author declared it. Still the PRIMARY event in every
+ * mode: it is the only presentation that carries the sample's own `fields` and
+ * the type-specific aliases in resolveField (an llm_input event resolves
+ * `user_input` from event.content; an mcp_exchange event does not). Keeping it
+ * in every mode is what makes the wider modes strict supersets — widening can
+ * reveal a match, never hide one.
  */
-function sampleToEvent(sample: CorpusSample): AgentEvent {
+function declaredEvent(sample: CorpusSample): AgentEvent {
   return {
     type: sample.eventType,
     content: sample.text,
@@ -53,30 +101,80 @@ function sampleToEvent(sample: CorpusSample): AgentEvent {
 }
 
 /**
+ * The events a sample is pushed through, primary first.
+ *
+ * Note what an eval needs that a rule-ID audit does not: the question here is
+ * "was this sample judged an attack at all", so the answer is the UNION over
+ * presentations, counted once per sample. A document that trips three rules on
+ * four shapes is one detection, not twelve — matching how scripts/ counts a
+ * false positive per sample rather than per shape.
+ */
+function sampleEvents(
+  sample: CorpusSample,
+  mode: EvalEventShapeMode
+): readonly AgentEvent[] {
+  const primary = declaredEvent(sample);
+  if (mode === 'legacy') return [primary];
+  if (mode === 'wide') return [primary, sourceAgnosticEvent(primary)];
+  return [primary, ...corpusShapes(sample.text).map((s) => s.event)];
+}
+
+/** scanSkill() is a separate engine entry point, not an event shape. */
+function usesSkillScan(mode: EvalEventShapeMode): boolean {
+  return mode === 'canonical';
+}
+
+/** Rule ids from the secondary presentations, via the safe public sync entry point. */
+function secondaryMatchIds(
+  engine: ATREngine,
+  sample: CorpusSample,
+  mode: EvalEventShapeMode
+): Map<string, number> {
+  const extra = new Map<string, number>();
+  for (const event of sampleEvents(sample, mode).slice(1)) {
+    for (const m of engine.evaluate(event)) {
+      extra.set(m.rule.id, Math.max(extra.get(m.rule.id) ?? 0, m.confidence));
+    }
+  }
+  if (usesSkillScan(mode)) {
+    for (const m of engine.scanSkill(sample.text)) {
+      extra.set(m.rule.id, Math.max(extra.get(m.rule.id) ?? 0, m.confidence));
+    }
+  }
+  return extra;
+}
+
+/**
  * Run a single sample through the engine (regex only) and measure results.
  * Catches engine errors so a single bad sample doesn't abort the entire eval.
  */
 function evaluateSampleRegex(
   engine: ATREngine,
-  sample: CorpusSample
+  sample: CorpusSample,
+  mode: EvalEventShapeMode
 ): SampleResult {
-  const event = sampleToEvent(sample);
-
   try {
     const start = performance.now();
-    const matches = engine.evaluate(event);
+    const byRule = new Map<string, number>();
+    for (const event of sampleEvents(sample, mode)) {
+      for (const m of engine.evaluate(event)) {
+        byRule.set(m.rule.id, Math.max(byRule.get(m.rule.id) ?? 0, m.confidence));
+      }
+    }
+    if (usesSkillScan(mode)) {
+      for (const m of engine.scanSkill(sample.text)) {
+        byRule.set(m.rule.id, Math.max(byRule.get(m.rule.id) ?? 0, m.confidence));
+      }
+    }
     const latencyMs = performance.now() - start;
-
-    const detected = matches.length > 0;
-    const topMatch = matches[0];
 
     return {
       id: sample.id,
       category: sample.category,
       expectedDetection: sample.expectedDetection,
-      actualDetection: detected,
-      matchedRules: matches.map((m) => m.rule.id),
-      confidence: topMatch?.confidence ?? 0,
+      actualDetection: byRule.size > 0,
+      matchedRules: [...byRule.keys()],
+      confidence: Math.max(0, ...byRule.values()),
       latencyMs,
       difficulty: sample.difficulty,
       tier: sample.tier,
@@ -98,27 +196,39 @@ function evaluateSampleRegex(
 
 /**
  * Run a single sample through the full pipeline (regex + embedding) and measure results.
+ *
+ * The Tier 2.5 embedding pass runs ONCE, on the declared event. It is not
+ * source-type filtered and it reads exactly one field (`content`), and every
+ * secondary presentation carries either the identical raw text or a JSON
+ * re-encoding of it — so encoding the sample four more times would multiply the
+ * runtime without changing a single verdict. The secondary presentations exist
+ * to fix rule ADMISSION, which is a Tier 2 (regex) concern, so they go through
+ * evaluate(): the public sync entry point, which still applies the lane gate and
+ * still withholds unconfirmed `confirm: embedding` rules outside the hunt lane.
  */
 async function evaluateSampleFull(
   engine: ATREngine,
-  sample: CorpusSample
+  sample: CorpusSample,
+  mode: EvalEventShapeMode
 ): Promise<SampleResult> {
-  const event = sampleToEvent(sample);
-
   try {
     const start = performance.now();
-    const { verdict } = await engine.evaluateWithVerdict(event);
+    const { verdict } = await engine.evaluateWithVerdict(declaredEvent(sample));
+    const byRule = new Map<string, number>(
+      verdict.matches.map((m) => [m.rule.id, m.confidence])
+    );
+    for (const [id, confidence] of secondaryMatchIds(engine, sample, mode)) {
+      byRule.set(id, Math.max(byRule.get(id) ?? 0, confidence));
+    }
     const latencyMs = performance.now() - start;
-
-    const detected = verdict.matchCount > 0;
 
     return {
       id: sample.id,
       category: sample.category,
       expectedDetection: sample.expectedDetection,
-      actualDetection: detected,
-      matchedRules: verdict.matches.map((m) => m.rule.id),
-      confidence: verdict.highestConfidence,
+      actualDetection: byRule.size > 0,
+      matchedRules: [...byRule.keys()],
+      confidence: Math.max(verdict.highestConfidence, 0, ...byRule.values()),
       latencyMs,
       difficulty: sample.difficulty,
       tier: sample.tier,
@@ -167,7 +277,9 @@ export async function runEval(config: EvalConfig): Promise<{
   corpusStats: ReturnType<typeof getCorpusStats>;
   tiersUsed: readonly string[];
   ruleQuality: RuleQualityReport;
+  eventShape: EvalEventShapeMode;
 }> {
+  const eventShape = config.eventShape ?? defaultEventShapeMode();
   const corpus = config.corpus ?? EVAL_CORPUS;
   const base = resolve(config.rulesDir, '..');
   const embeddingsPath = config.embeddingsPath ?? join(base, 'data', 'attack-embeddings.json');
@@ -201,8 +313,8 @@ export async function runEval(config: EvalConfig): Promise<{
 
   for (const sample of corpus) {
     const result = useFullPipeline
-      ? await evaluateSampleFull(engine, sample)
-      : evaluateSampleRegex(engine, sample);
+      ? await evaluateSampleFull(engine, sample, eventShape)
+      : evaluateSampleRegex(engine, sample, eventShape);
     results.push(result);
   }
 
@@ -230,11 +342,12 @@ export async function runEval(config: EvalConfig): Promise<{
       ruleCount,
       engine: 'ATREngine',
       tiers: tiersUsed,
+      eventShape,
     };
     writeFileSync(config.outputPath, JSON.stringify(output, null, 2));
   }
 
-  return { report, regression, corpusStats, tiersUsed, ruleQuality };
+  return { report, regression, corpusStats, tiersUsed, ruleQuality, eventShape };
 }
 
 // ---------------------------------------------------------------------------
@@ -256,7 +369,7 @@ export async function runEvalCLI(): Promise<void> {
 
   console.log('\n=== ATR Evaluation Harness ===\n');
 
-  const { report, regression, corpusStats, tiersUsed, ruleQuality } = await runEval({
+  const { report, regression, corpusStats, tiersUsed, ruleQuality, eventShape } = await runEval({
     rulesDir,
     outputPath,
   });
@@ -265,6 +378,7 @@ export async function runEvalCLI(): Promise<void> {
   console.log(`Corpus: ${corpusStats.total} samples (${corpusStats.attacks} attacks, ${corpusStats.benign} benign)`);
   console.log(`Categories: ${Object.keys(corpusStats.byCategory).join(', ')}`);
   console.log(`Tiers: ${tiersUsed.join(' + ')}`);
+  console.log(`Event shape: ${eventShape}`);
 
   // Overall metrics
   console.log(`\n--- Overall ---`);
@@ -364,7 +478,11 @@ export async function runEvalCLI(): Promise<void> {
   const { measurementPath } = writeMeasurementFromEvalReport(report, {
     source: 'atr-self-test',
     source_version: 'internal',
-    notes: "Engine regression eval against the union of every rule's own test_cases.",
+    notes:
+      "Engine regression eval against the union of every rule's own test_cases. " +
+      `Event shape: ${eventShape} (see src/eval/eval-harness.ts EvalEventShapeMode — ` +
+      'the mode decides how many rules the engine admits per sample, so a number ' +
+      'measured under one mode is not comparable with a number measured under another).',
   });
   console.log(`Measurement: ${measurementPath}`);
   console.log('Done.\n');
